requirements.txt
# Core CrewAI dependencies
crewai>=0.70.1
crewai-tools>=0.12.1

# OpenRouter and LLM integration
litellm>=1.48.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Terminal UI and formatting
rich>=13.7.0
click>=8.1.7

# Web scraping and search tools (for custom tools)
requests>=2.31.0
beautifulsoup4>=4.12.3

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Voice capabilities (HW4)
openai>=1.0.0  # Whisper STT API
kokoro-onnx>=0.3.0  # Kokoro TTS
sounddevice>=0.4.6  # Microphone recording
soundfile>=0.12.1  # Audio file handling
pydub>=0.25.1  # Audio processing